Write correctly composed BVH rotations
One interesting complexity is that the specified channel ordering in
the BVH format describes the composition order, which is the reverse
of the somewhat more intuitive application order that ASF/AMC uses.
That is, if you want to apply an X rotation, then a Y rotation,
followed by a Z rotation, you write Zrotation Xrotation Yrotation.
This seems to be a common practice in mocap formats.

Also removed a lot of debugging code.